OmopConstructor: Build Tables in the OMOP Common Data Model

Provides functionality to construct standardised tables from health care data formatted according to the Observational Medical Outcomes Partnership (OMOP) Common Data Model. The package includes tools to build key tables such as observation period and drug era, among others.

See also

Author

Maintainer: Martí Català marti.catalasabate@ndorms.ox.ac.uk (ORCID)

Other contributors:

Elin Rowlands elin.rowlands@ndorms.ox.ac.uk (ORCID) [contributor]

Cecilia Campanile cecilia.campanile@ndorms.ox.ac.uk (ORCID) [contributor]